Only subscribe to breakoutrooms if capable. Resolves a bug where ecs disabled flag causes breakoutroom initialization to fail (#5978)
* Only subscribe to breakoutrooms if capable. Resolves a bug when ecs is disabled for breakoutrooms

* stable changelog

* beta changelog

* Add checks to ensure capabiltiies object is available

+++ packages/calling-stateful-client/src/CallSubscriber.ts
@@ -62,7 +62,7 @@ export class CallSubscriber {
 
   private _capabilitiesSubscriber: CapabilitiesSubscriber;
   private _spotlightSubscriber: SpotlightSubscriber;
-  private _breakoutRoomsSubscriber: BreakoutRoomsSubscriber;
+  private _breakoutRoomsSubscriber?: BreakoutRoomsSubscriber;
 
   private _togetherModeSubscriber: TogetherModeSubscriber;
   private _mediaAccessSubscriber: MediaAccessSubscriber;
@@ -120,13 +120,15 @@ export class CallSubscriber {
       this._call.feature(Features.Spotlight)
     );
 
-    // Clear assigned breakout room closed notification for this call.
-    this._context.deleteLatestNotification('assignedBreakoutRoomClosed');
-    this._breakoutRoomsSubscriber = new BreakoutRoomsSubscriber(
-      this._callIdRef,
-      this._context,
-      this._call.feature(Features.BreakoutRooms)
-    );
+    if (this._call.feature(Features.Capabilities) && this._call.feature(Features.Capabilities).capabilities?.joinBreakoutRooms.isPresent) {
+      // Clear assigned breakout room closed notification for this call.
+      this._context.deleteLatestNotification('assignedBreakoutRoomClosed');
+      this._breakoutRoomsSubscriber = new BreakoutRoomsSubscriber(
+        this._callIdRef,
+        this._context,
+        this._call.feature(Features.BreakoutRooms)
+      );
+    }
 
     this._togetherModeSubscriber = new TogetherModeSubscriber(
       this._callIdRef,
@@ -254,8 +256,9 @@ export class CallSubscriber {
     this._capabilitiesSubscriber.unsubscribe();
     this._reactionSubscriber?.unsubscribe();
     this._spotlightSubscriber.unsubscribe();
-    this._breakoutRoomsSubscriber.unsubscribe();
-
+    if (this._call.feature(Features.Capabilities) && this._call.feature(Features.Capabilities).capabilities?.joinBreakoutRooms.isPresent && this._breakoutRoomsSubscriber) {
+      this._breakoutRoomsSubscriber.unsubscribe();
+    }
     this._togetherModeSubscriber.unsubscribe();
     this._mediaAccessSubscriber.unsubscribe();
   };
